backend/requirements-dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements-dev.in
#
appdirs==1.4.4
    # via black
attrs==20.3.0
    # via pytest
backcall==0.2.0
    # via ipython
black==20.8b1
    # via -r requirements-dev.in
click==7.1.2
    # via black
coverage==5.5
    # via pytest-cov
decorator==5.0.7
    # via ipython
factory-boy==3.2.0
    # via -r requirements-dev.in
faker==8.1.1
    # via factory-boy
flake8==3.9.1
    # via -r requirements-dev.in
iniconfig==1.1.1
    # via pytest
ipython-genutils==0.2.0
    # via traitlets
ipython==7.22.0
    # via -r requirements-dev.in
isort==5.8.0
    # via -r requirements-dev.in
jedi==0.18.0
    # via ipython
mccabe==0.6.1
    # via flake8
mypy-extensions==0.4.3
    # via black
packaging==20.9
    # via pytest
parso==0.8.2
    # via jedi
pathspec==0.8.1
    # via black
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pluggy==0.13.1
    # via pytest
prompt-toolkit==3.0.18
    # via ipython
ptyprocess==0.7.0
    # via pexpect
py==1.10.0
    # via pytest
pycodestyle==2.7.0
    # via flake8
pyflakes==2.3.1
    # via flake8
pygments==2.8.1
    # via ipython
pyparsing==2.4.7
    # via packaging
pytest-cov==2.11.1
    # via -r requirements-dev.in
pytest-django==4.2.0
    # via -r requirements-dev.in
pytest==6.2.3
    # via
    #   -r requirements-dev.in
    #   pytest-cov
    #   pytest-django
python-dateutil==2.8.1
    # via faker
regex==2021.4.4
    # via black
six==1.15.0
    # via python-dateutil
text-unidecode==1.3
    # via faker
toml==0.10.2
    # via
    #   black
    #   pytest
traitlets==5.0.5
    # via ipython
typed-ast==1.4.3
    # via black
typing-extensions==3.7.4.3
    # via black
wcwidth==0.2.5
    # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
